Add kotlinx.rpc
I want to add filtering and pagination to to the REST endpoint
/index-surveys and it seemed like an opportunity to introduce
kotlinx.rpc in order to reduce friction.

This can be seen as a controversial choice. REST is very common and
kRPC is not based on a standard that is implemented by other
programming languages.

You can test it in the browser console:

```
const {testRpcClient} = await import("/src/index.tsx")
testRpcClient()
```

It should log something like

```
Object { records: {…}, numberOfMatches: 0 }
```